requirements.txt
streamlit
selenium
beautifulsoup4
pandas
google-generativeai
